Reuse exact copies of runs
If uploading splits which have already been uploaded, back the run with the same RunFile so we're not double storing
the data. The new upload will still be represented as its own set of splits with its own URL and owner and all that;
this change should be entirely unnoticeable.

Closes #13, fixes #66