Auto associate runs' runners with SRDC runs'
On run creation/update, in a background job: If a run has no assigned user and does have an assigned srdc_id, check the
speedrun.com API for the runner, and check that runner's associated SRDC profile for an associated Twitch account. If
that Twitch account has an associated splits.io account, automatically tie the run to that account.

tldr; complete the splits.io <-> twitch <-> speedrun.com connection