Fix game/category route variable names.
Using `resources` is nice to start, but when you want to name the params the same thing (:game, :category, :run, etc.)
for every occurrence it's hard to wrangle. And it takes more effort to decipher a resources call if you're not
implementing full CRUD in the simplest fashion.

So we are moving to setting paths more manually like this. Last commit that touched this forgot to clean up some
controllers though, so this fixes #250. Also removes some Rails 5 deprecation warnings from tests.