Fix broken game links
Originally I was going to remove linking to games based on ID, but it's
needed if we want to support permalink redirectors without refactoring
`to_param` into something at the controller level.

So this lets you link to a game with splits.io/games/sms OR
splits.io/games/15. This is a bit dirty ("what about a game with a
number for a name?") but is a decent enough solution until we move
`to_param` logic into the controller level.