Don't begin segment history graph with nulls
Addresses an issue in #271 comments

In short: If the segment history graph starts with a bunch of nulls,
don't display them. This is a bit tricky since all segments reside on
the same graph, so a graph showing one segment could actually still show
some nulls at the beginning if another segment that can be shown on that
graph doesn't have nulls in those slots; but this addresses the big
issue, which is that some people have splits from before LiveSplit
started recording segment history.