Fix an apparent (but not really) rounding error
When we display time differences between runs (like on the rivals page)
as raw differences like "-00:07:34", that time is sometimes not actually
the difference between the two runs' displayed times.

The reason:

My run takes 1000000ms. Yours takes 1234567ms. The displayed times are
00:16:40 and 00:20:34 respectively. When calculating the differences we
use the raw ms then truncate, so the displyed difference is

    1234567ms - 1000000ms = 234567ms = 00:03:54

But napkin math shows that

    00:20:34 - 00:16:40 = 00:03:53

and not 00:03:54.

So the displayed times (the more precise ones) look as if they're
subject to rounding errors, when actually no rounding takes place on the
website (only truncating for display).

To fix this, displayed differences are now calculated from displayed
times (i.e. pre-truncated times), not from raw times then truncated
later.