Move segment history graphs onto the run page (#321)
* Move segment history graphs onto the run page

Moves that multi-line segment history graph up into one graph per
segment, and puts them right inside the segment entries on the run show
page.

This is kinda the first step of rolling the stats and compare pages into
the standard run page (related: #263). Historically we've kept a lot of
this stuff on the stats page because we need to look at history to
generate it, and looking at history is expensive. But that's where all
the useful stuff is, so I don't think we should keep it out of the
limelight. We can figure out how to scale it as we go.

* Add a comment

* Remove color_styles

* Move other graph JS into graphs folder

* Fix History title not appearing on gametime runs

* Fix turbolinks naming conflict