Rudimentary browser source layout for races (#659)
Adds a v1 browser source friendly layout for the race page. This is like
the current race page, but:

1. Elements are resizeable by the user
2. Text is editable by the user
3. Elements have a black background with red borders and extra padding

The next item is to add a feed to this page with all the things we've
talked about that could fuel interesting commentary during a race, but
that part has some backend code and is worth its own branch.