Refactor Twitch follows
Refactors Twitch follows to be relationships between TwitchUsers, not
Users. At the time TwitchUserFollows were created, TwitchUsers did not
yet exist so the follow relationship operated on Users directly.

Now that we operate on TwitchUsers, we get some additional behavior
perks for free, like cascading deletion of follows when unlinking a
Twitch account. As a bonus we get the more idiomatic syntax of
`user.twitch.follows` rather than `user.twitch_follows`.

This also fixes some sync bugs with Twitch follows and synchronization
along the way, which were ultimately causing the front page feed to be
empty when it shouldn't be.